HPCC-29313 ECL Watch v9 fix Dynamic ESDL binding selection error
